refactor(core): move the stream-network construction where every layer can reach it
