+++ hydromodpy/core/stream_network.py
